+++ index.html
@@ -463,7 +463,7 @@
     </template>
 
     <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-    <script src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
 
   </body>
 </html>
